requirements.txt
asgiref==3.8.1
celery==5.3.6
coverage==7.8.0
Django==4.2
django-cors-headers==4.7.0
djangorestframework==3.16.0
drf-yasg==1.21.10
flower==2.0.1
gunicorn==23.0.0
inflection==0.5.1
iniconfig==2.1.0
packaging==25.0
pillow==11.2.1
pluggy==1.6.0
pytest==8.3.5
pytest-django==4.11.1
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
redis==5.0.1
sqlparse==0.5.3
uritemplate==4.1.1
